Add a workaround for dev on whitelist to the existing snapshot jobs

+++ snapshot/snapshot.sh
@@ -58,19 +58,24 @@ tmp_pass=$4
 tmp_host=dbutil.drupal.org
 tmp_args="-h${tmp_host} -u${tmp_user} -p${tmp_pass} ${tmp_db}"
 
-clear_tmp
-
-# Make a copy of live.
-mysqldump -h$db_host -u$db_user -p$db_pass --single-transaction --quick $db_name 2> "${WORKSPACE}/mysqldump-errors.txt" | gzip > "${WORKSPACE}/tmp.mysql.gz"
-# Check for errors.
-if [ -s "${WORKSPACE}/mysqldump-errors.txt" ]; then
+# Work around clearing out a copy of the database when using the whitelist for
+# dev (which populates drupal_sanitize for us). Prevents doing a duplicate
+# mysqldump from the live site when using the whitelist.
+if [ ${db_name} != 'drupal' ]; then
+  clear_tmp
+  # Make a copy of live.
+  mysqldump -h$db_host -u$db_user -p$db_pass --single-transaction --quick $db_name 2> "${WORKSPACE}/mysqldump-errors.txt" | gzip > "${WORKSPACE}/tmp.mysql.gz"
+  # Check for errors.
+  if [ -s "${WORKSPACE}/mysqldump-errors.txt" ]; then
+    rm "${WORKSPACE}/tmp.mysql.gz"
+    cat "${WORKSPACE}/mysqldump-errors.txt"
+    exit 1
+  fi
+  # Copy live to tmp database.
+  gunzip < "${WORKSPACE}/tmp.mysql.gz" | sed -e 's/^) ENGINE=[^ ]*/) ENGINE=MyISAM/' | mysql -o ${tmp_args}
   rm "${WORKSPACE}/tmp.mysql.gz"
-  cat "${WORKSPACE}/mysqldump-errors.txt"
-  exit 1
 fi
-# Copy live to tmp database.
-gunzip < "${WORKSPACE}/tmp.mysql.gz" | sed -e 's/^) ENGINE=[^ ]*/) ENGINE=MyISAM/' | mysql -o ${tmp_args}
-rm "${WORKSPACE}/tmp.mysql.gz"
+
 
 # Save a copy of the schema.
 mysqldump --single-transaction --quick ${tmp_args} -d --compact --skip-opt > "${WORKSPACE}/schema.mysql"
